Has a working "send to miaou".
However, it will have to be changed. I have to use the background page
as the page communicating with the native app.

So that we end up with these communications:

1. popup -> background -> content_script to send encrypted messages
2. content_script -> background to decrypt encrypted messages it found
   via MutationObserver
3. popup -> background when a popup opens, to get the list of decrypted
   messages the background stored (that it got from 2.)

The problem is that popup is short-lived, while background is
long-lived. So background can get new messages coming in miaou and store
them, and then communicate them to popup when necessary.