Fixes the verify to show up on loading
Before this change, if you get on the discussion page right away, you're
on the loading page, so the elements don't exist at this point.

However, there are cases where you can get in this module and still not
be on the discussion page... so we just give up after 3 times.

In the worst case (loading > 3 seconds), the user can come back to the
home page, back to the mail, and the verify button will be there.